Remove esm dependency
