+++ package.json
@@ -44,7 +44,6 @@
     "@kazupon/vue-i18n-loader": "^0.3.0",
     "acorn": "^5.6.2",
     "cookie": "^0.3.1",
-    "esm": "^3.0.76",
     "js-cookie": "^2.2.0",
     "vue-i18n": "^8.0.0",
     "vue-i18n-extensions": "^0.2.0"
